% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normlize.R
\name{logNorm}
\alias{logNorm}
\alias{logNorm.matrix}
\alias{logNorm.dgCMatrix}
\title{Normalize and log-transform raw data}
\usage{
logNorm(object, ...)

\method{logNorm}{matrix}(object, scale.factor = 10000, ...)

\method{logNorm}{dgCMatrix}(object, scale.factor = 10000, ...)
}
\arguments{
\item{object}{Matrix with the raw data.}

\item{...}{Arguments passed to other methods.}

\item{scale.factor}{The total count after normalization. Default is 10000.}
}
\description{
Normalize each cell by relative counts, then natural log-transform the data.
}
