% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normlize.R
\name{relativeCounts}
\alias{relativeCounts}
\alias{relativeCounts.matrix}
\alias{relativeCounts.dgCMatrix}
\title{Normalize data using relative counts}
\usage{
relativeCounts(object, ...)

\method{relativeCounts}{matrix}(object, scale.factor = 10000, ...)

\method{relativeCounts}{dgCMatrix}(object, scale.factor = 10000, ...)
}
\arguments{
\item{object}{Matrix with the raw data.}

\item{...}{Arguments passed to other methods.}

\item{scale.factor}{The total count after normalization. Default is 10000.}
}
\description{
Normalize each cell by total counts over all genes, so that every cell has
the same total count after normalization.
}
