% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowExpMean.R
\name{rowExpMean}
\alias{rowExpMean}
\alias{rowExpMean.numeric}
\alias{rowExpMean.CsparseMatrix}
\alias{rowExpMean.matrix}
\title{Calculate the row means of logged values}
\usage{
rowExpMean(object, ...)

\method{rowExpMean}{numeric}(object, log1p = FALSE, ...)

\method{rowExpMean}{CsparseMatrix}(object, log1p = FALSE, ...)

\method{rowExpMean}{matrix}(object, log1p = FALSE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods.}

\item{log1p}{Logical, whether to transform the result into log-space. Default
is \code{FALSE}.}
}
\value{
A numeric vector containing the row means of \code{object}. If \code{object} is a 1-D
array, return a numeric scalar.
}
\description{
Calculate the row means of logged values in non-log space.
}
