% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChange.R
\name{selectDE}
\alias{selectDE}
\title{Select genes from differential expression analysis}
\usage{
selectDE(
  fc.results,
  min.diff = 0.1,
  min.mean.exp = 0,
  min.pct = 0.01,
  min.diff.pct = -Inf,
  min.cells.feature = 3,
  only.pos = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fc.results}{A \code{\link[=data.frame]{data.frame()}} from \code{\link[=foldChange]{foldChange()}} containing:
\itemize{
\item a first column representing fold change (\code{avg_log2FC}, \code{avg_diff}, etc.)
\item \code{mean.1}, \code{mean.2}
\item \code{ncells.1}, \code{ncells.2}
\item \code{pct.1}, \code{pct.2}
}}

\item{min.diff}{Minimum absolute difference or log fold change) required.
Must be \verb{>= 0}. Default is 0.1.}

\item{min.mean.exp}{Minimum mean expression required in at least one group.}

\item{min.pct}{Select genes that are detected in a minimum fraction of
\code{min.pct} cells in either of the two populations.}

\item{min.diff.pct}{Minimum difference in detection fraction between groups.}

\item{min.cells.feature}{Minimum number of cells expressing the feature
in at least one group.}

\item{only.pos}{Logical. If \code{TRUE}, only retain features with positive
difference.}

\item{verbose}{Logical, whether to print the progress information to console.}
}
\value{
A filtered \code{\link[=data.frame]{data.frame()}} containing only features that pass all thresholds.
}
\description{
Filter fold-change results (typically produced by \code{\link[=foldChange]{foldChange()}}) according
to multiple expression and detection thresholds.
}
\seealso{
\code{\link[=foldChange]{foldChange()}}
}
